Upgrade to Kotlin 1.4-M3
