+++ buildSrc/build.gradle.kts
@@ -7,10 +7,12 @@ plugins {
 repositories {
     jcenter()
     gradlePluginPortal()
+    maven ("https://dl.bintray.com/kotlin/kotlin-eap")
+    maven ("https://kotlin.bintray.com/kotlinx")
 }
 
 dependencies {
-    api(kotlin("gradle-plugin", version = "1.3.71"))
+    api(kotlin("gradle-plugin", version = "1.4-M3"))
     api("org.jetbrains.dokka:dokka-gradle-plugin:0.9.17")
     api("com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4")
     api("gradle.plugin.org.jetbrains.intellij.plugins:gradle-intellij-plugin:0.4.10")
