+++ integration-test/build.gradle.kts
@@ -5,6 +5,8 @@ plugins {
 
 repositories {
   jcenter()
+  maven ("https://dl.bintray.com/kotlin/kotlin-eap")
+  maven ("https://kotlin.bintray.com/kotlinx")
 }
 
 kotlin {
