+++ integration-test/buildSrc/build.gradle.kts
@@ -1,7 +1,9 @@
 repositories {
   jcenter()
+  maven ("https://dl.bintray.com/kotlin/kotlin-eap")
+  maven ("https://kotlin.bintray.com/kotlinx")
 }
 
 dependencies {
-  implementation(kotlin("gradle-plugin", version = "1.3.72"))
+  implementation(kotlin("gradle-plugin", version = "1.4-M3"))
 }
