+++ integration-test/gradle.properties
@@ -1,1 +1,1 @@
-org.gradle.jvmArgs=-Xmx3G
+org.gradle.jvmargs=-Xmx3G
