+++ integration-test/settings.gradle.kts
@@ -18,6 +18,11 @@ includeBuild("../") {
 }
 
 pluginManagement {
+  repositories {
+    jcenter()
+    maven ("https://dl.bintray.com/kotlin/kotlin-eap")
+    maven ("https://kotlin.bintray.com/kotlinx")
+  }
   resolutionStrategy.eachPlugin {
     if (requested.id.id == "org.spekframework.spek2.multiplatform") {
       useModule("org.spekframework.spek2:spek-gradle-plugin:0.1.0")
